Add VERSION, start at 10.00
Our source code will originate from fecon235 v5.18.0312
which is date-based, so circa 2018-03-12.
The history of that repository goes back to the year 2014:
see https://github.com/rsvp/fecon235 for complete details.

The rationale for fecon236 is given in:
"Moving to Python 3 > 2020-01-01"
https://github.com/rsvp/fecon235/issues/9

The source code repo should be separate from bulky size
of Jupyter notebooks, thus they are expected to
remain in the fecon235 repository and be developed further there.